### 2025-04-23 #8098 @danielskatz
👋 @ChristopheBerthod - thanks for your submission.

We don't currently have an editor, so I will put this on our waiting list, but hopefully it won't be long. In the meantime, you could change the place in the .md file where you say "See [@ref]" to "See @ref", following the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html). You can the use the command `@editorialbot generate pdf` after to make a new PDF. editorialbot commands need to be the first entry in a new comment.
