### 2025-06-02 #8329 @editorialbot
[REVIEW]: Piecewise: Flexible piecewise functions for fast integral transforms in Julia
**Submitting author:** @ChristopheBerthod (<a href="http://orcid.org/0000-0002-0787-008X">Christophe Berthod</a>)
**Repository:** https://github.com/ChristopheBerthod/Piecewise.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
