### 2025-07-24 #8329 @gkthiruvathukal
@ChristopheBerthod Yes, please try to distill the feedback from @dawbarton and give a point by point response. Based on my reading, the feedback is aimed at sharpening your submission with an eye to acceptance. As an editor, that is my goal.
