+++ dev/assets/documenter.js
@@ -104,7 +104,7 @@ $(document).on(
         .prop("title", articleToggleTitle);
       parent.siblings("section").slideToggle();
     });
-  }
+  },
 );
 
 $(document).on("click", ".docs-article-toggle-button", function (event) {
@@ -239,6 +239,20 @@ $(document).ready(function () {
       type: "click",
       noToggleAnimation: true,
     });
+
+    setTimeout(function () {
+      if (window.location.hash) {
+        const targetId = window.location.hash.substring(1);
+        const targetElement = document.getElementById(targetId);
+
+        if (targetElement) {
+          targetElement.scrollIntoView({
+            behavior: "smooth",
+            block: "center",
+          });
+        }
+      }
+    }, 100);
   }
 });
 
@@ -312,7 +326,7 @@ update_search
 
 function worker_function(documenterSearchIndex, documenterBaseURL, filters) {
   importScripts(
-    "https://cdn.jsdelivr.net/npm/minisearch@6.1.0/dist/umd/index.min.js"
+    "https://cdn.jsdelivr.net/npm/minisearch@6.1.0/dist/umd/index.min.js",
   );
 
   let data = documenterSearchIndex.map((x, key) => {
@@ -523,8 +537,8 @@ function worker_function(documenterSearchIndex, documenterBaseURL, filters) {
             Math.max(textindex.index - 100, 0),
             Math.min(
               textindex.index + querystring.length + 100,
-              result.text.length
-            )
+              result.text.length,
+            ),
           )
         : ""; // cut-off text before and after from the match
 
@@ -534,7 +548,7 @@ function worker_function(documenterSearchIndex, documenterBaseURL, filters) {
       ? "..." +
         text.replace(
           new RegExp(`${escape(searchstring)}`, "i"), // For first occurrence
-          '<span class="search-result-highlight py-1">$&</span>'
+          '<span class="search-result-highlight py-1">$&</span>',
         ) +
         "..."
       : ""; // highlights the match
@@ -547,7 +561,7 @@ function worker_function(documenterSearchIndex, documenterBaseURL, filters) {
     // We encode the full url to escape some special characters which can lead to broken links
     let result_div = `
         <a href="${encodeURI(
-          documenterBaseURL + "/" + result.location
+          documenterBaseURL + "/" + result.location,
         )}" class="search-result-link w-100 is-flex is-flex-direction-column gap-2 px-4 py-2">
           <div class="w-100 is-flex is-flex-wrap-wrap is-justify-content-space-between is-align-items-flex-start">
             <div class="search-result-title has-text-weight-bold ${
@@ -647,6 +661,42 @@ function runSearchMainCode() {
   // Which filter is currently selected
   var selected_filter = "";
 
+  document.addEventListener("reset-filter", function () {
+    selected_filter = "";
+    update_search();
+  });
+
+  //update the url with search query
+  function updateSearchURL(query) {
+    const url = new URL(window.location);
+
+    if (query && query.trim() !== "") {
+      url.searchParams.set("q", query);
+    } else {
+      // remove the 'q' param if it exists
+      if (url.searchParams.has("q")) {
+        url.searchParams.delete("q");
+      }
+    }
+
+    // Add or remove the filter parameter based on selected_filter
+    if (selected_filter && selected_filter.trim() !== "") {
+      url.searchParams.set("filter", selected_filter);
+    } else {
+      // remove the 'filter' param if it exists
+      if (url.searchParams.has("filter")) {
+        url.searchParams.delete("filter");
+      }
+    }
+
+    // Only update history if there are parameters, otherwise use the base URL
+    if (url.search) {
+      window.history.replaceState({}, "", url);
[…]
