+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"lorentz.html#PiecewiseLorentz","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"","category":"section"},{"location":"lorentz.html#Installation","page":"PiecewiseLorentz","title":"Installation","text":"","category":"section"},{"location":"lorentz.html","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"The package won't work without Piecewise. Both can be installed with Pkg.add.","category":"page"},{"location":"lorentz.html","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"using Pkg; Pkg.add(\"Piecewise\"); Pkg.add(\"PiecewiseLorentz\")","category":"page"},{"location":"lorentz.html#Introduction","page":"PiecewiseLorentz","title":"Introduction","text":"","category":"section"},{"location":"lorentz.html","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"The order-m Lorentz transform of a function f(x) of a real variable x is defined as","category":"page"},{"location":"lorentz.html","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"left(L^mcirc fright)(y z) = int_-infty^inftydxf(x)\nleftfrac-mathrmImzpi(y-mathrmRez-x)^2+(mathrmImz)^2right^m","category":"page"},{"location":"lorentz.html","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"where m is a positive integer, y is a real number, and z is a complex number with strictly negative imaginary part. The name Lorentz transform was chosen, because the function in square brackets is a Lorentzian of half-width mathrmImz centered at x=y-mathrmRez.","category":"page"},{"location":"lorentz.html","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"The module PiecewiseLorentz adds to the Formulas defined in the module Piecewise the primitive functions corresponding to the order-m Lorentz kernel leftfrac-mathrmImzpi(y-mathrmRez-x)^2+(mathrmImz)^2right^m, enabling the fast Lorentz transform of PiecewiseFunction objects using these formulas.","category":"page"},{"location":"lorentz.html","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"warning: Warning\nThe present version only implements m=123 for the formulas POLY, LOG, PLS, and XLOG. Note that the m=1 Lorentz transform is equivalent to a Hilbert transform:left(L^1circ fright)(y z) = -frac1pimathrmIm\nleft(Hcirc fright)(y-z)","category":"page"},{"location":"lorentz.html","page":"PiecewiseLorentz","title":"PiecewiseLorentz","text":"The method lorentz_transform return the Lorentz transform of a PiecewiseFunction object, as calculated using the definition above. The transform has algebraic tails at large y, that may suffer from numerical errors. A numerically more stable moment expansion is therefore used for large values of y-mathrmRez. Both the definition and the moment expansion are implemented in a LorentzTransform object created a
[…]
