### 2025-07-24 be8ceaa
joss paper
