+++ joss/paper.md
@@ -21,15 +21,15 @@ A piecewise function of a real variable $x$ returns a value computed from a rule
 
 # Statement of need
 
-The interpolation problem, which consists in constructing a continuous function out of discrete data, is ubiquitous in many areas of science and technology. This problem has been traditionally solved by means of global or piecewise polynomial functions [@Numerical-Methods-2021; @Bhagavan-2024; @Interpolations.jl]. The various interpolation schemes differ by the order of the polynomials, the choice of the sampling points when this choice is possible, and the additional conditions required when the solution is not uniquely determined by the data. Beside drawing a smooth curve through points in a graph, one important use of interpolations is for constructing a cheap but accurate approximation of a computer-intensive function. If the latter function presents critical points like discontinuities or singularities, all polynomial interpolations fail in the neighborhood of these points, due to the absence of a convergent Taylor series. When the underlying function has critical points and accuracy is an issue, there is a need for piecewise interpolation schemes that are based on nonanalytic functions rather than polynomials.
+The interpolation problem, which consists in constructing a continuous function out of discrete data, is ubiquitous in many areas of science and technology. This problem has been traditionally solved by means of global or piecewise polynomial functions [@Numerical-Methods-2021; @Bhagavan-2024; @Interpolations.jl]. The various interpolation schemes differ by the order of the polynomials, the choice of the sampling points when this choice is possible, and the additional conditions required when the solution is not uniquely determined by the data. Beside drawing a smooth curve through points in a graph, one important use of interpolations is for constructing a cheap but accurate approximation of a computer-intensive function. Powerful tools have been developed for smooth functions that are well approximated by polynomial interpolants using Chebyshev points [@Chebfun; @ApproxFun.jl-2014]. However, if the function presents critical points like discontinuities or singularities, all polynomial interpolations fail in the neighborhood of these points, due to the absence of a convergent Taylor series. When the underlying function has critical points and accuracy is an issue, there is a need for piecewise interpolation schemes that are based on nonanalytic functions rather than polynomials.
 
[…]
