+++ joss/paper.bib
@@ -69,3 +69,21 @@
   pages     = {100002},
   doi       = {10.1016/j.newton.2024.100002}
 }
+
+@book{Chebfun,
+  editor       = {Driscoll, T. A. and Hale, N. and Trefethen, L. N.},
+  title        = {{Chebfun Guide}},
+  publisher    = {Pafnuty Publications},
+  year         = {2014},
+  address      = {Oxford},
+  howpublished = {\url{https://www.chebfun.org}}
+}
+
+@inproceedings{ApproxFun.jl-2014,
+  title     = {A practical framework for infinite-dimensional linear algebra},
+  author    = {Olver, Sheehan and Townsend, Alex},
+  booktitle = {Proceedings of the 1st Workshop for High Performance Technical Computing in Dynamic Languages -- {HPTCDL} `14},
+  year      = {2014},
+  publisher = {{IEEE}},
+  doi       = {10.1109/HPTCDL.2014.10}
+}
