### 2025-07-25 #8329 @ranocha
Thanks for your contribution! I started to review the paper and the software, but could not install the software (and opened an issue in your repository). Thus, I have to postpone the software review to later (which likely means the end of August or September).

The paper is well-written. I just have one question: Do you know how your approach relates to other approximation algorithms, such as rational approximations (available, for example, in https://github.com/complexvariables/RationalFunctionApproximation.jl)? Rational approximations are often described as working well in the presence of singularities.
