+++ docs/Project.toml
@@ -12,7 +12,7 @@ PiecewiseLorentz = {path = "../PiecewiseLorentz/"}
 
 [compat]
 Documenter = "1.8"
-Piecewise = "0.1.0"
-PiecewiseHilbert = "0.1.0"
-PiecewiseLorentz = "0.1.0"
+Piecewise = "0.1.1"
+PiecewiseHilbert = "0.1.1"
+PiecewiseLorentz = "0.1.1"
 Plots = "1.40.9"
