+++ PiecewiseHilbert/Project.toml
@@ -1,7 +1,7 @@
 name = "PiecewiseHilbert"
 uuid = "fb75a84c-a7ae-4011-94d8-abc4bb6dbff9"
 authors = ["Christophe Berthod <christophe.berthod@unige.ch>"]
-version = "0.1.0"
+version = "0.1.1"
 
 [deps]
 HypergeometricFunctions = "34004b35-14d8-5ef3-9330-4cdb6864b03a"
@@ -12,7 +12,7 @@ PolyLog = "85e3b03c-9856-11eb-0374-4dc1f8670e7f"
 HypergeometricFunctions = "0.3.27"
 Piecewise = "0.1.0"
 PolyLog = "2.5.1"
-Test = "1.11.0"
+Test = "1.10"
 julia = "1.10"
 
 [extras]
