### 2025-07-25 ChristopheBerthod/Piecewise.jl#9 @ChristopheBerthod
Thanks! Revised the problematic Project.toml files. The installation should now work with Julia 1.10.9.
