### 2025-07-25 ChristopheBerthod/Piecewise.jl#9 @ranocha
Thanks
