### 2025-07-25 ChristopheBerthod/Piecewise.jl#13 @ranocha
Tests fail
This is a part of https://github.com/openjournals/joss-reviews/issues/8329

Thanks for your package. Installation works on Julia 1.10.9, but tests fail, e.g.

```julia
(2025_07_Piecewise_jl) pkg> test Piecewise
     Testing Piecewise
      Status `.../Project.toml`
  [34004b35] HypergeometricFunctions v0.3.28
  [2fda8390] LsqFit v0.15.1
  [f03ddd90] Piecewise v0.1.1
  [de0858da] Printf
  [8dfed614] Test
      Status `.../Manifest.toml`
  [47edcb42] ADTypes v1.15.0
  [79e6a3ab] Adapt v4.3.0
  [66dad0bd] AliasTables v1.1.3
  [4fba245c] ArrayInterface v7.19.0
  [bbf7d656] CommonSubexpressions v0.3.1
  [34da2185] Compat v4.17.0
  [187b0558] ConstructionBase v1.6.0
  [9a962f9c] DataAPI v1.16.0
  [864edb3b] DataStructures v0.18.22
  [163ba53b] DiffResults v1.1.0
  [b552c78f] DiffRules v1.15.1
  [a0c0ee7d] DifferentiationInterface v0.7.3
  [31c24e10] Distributions v0.25.120
  [ffbed154] DocStringExtensions v0.9.5
  [1a297f60] FillArrays v1.13.0
  [6a86dc24] FiniteDiff v2.27.0
  [f6369f11] ForwardDiff v1.0.1
  [34004b35] HypergeometricFunctions v0.3.28
  [92d709cd] IrrationalConstants v0.2.4
  [692b3bcd] JLLWrappers v1.7.1
  [2ab3a3ac] LogExpFunctions v0.3.29
  [2fda8390] LsqFit v0.15.1
  [1914dd2f] MacroTools v0.5.16
  [e1d29d7a] Missings v1.2.0
  [d41bc354] NLSolversBase v7.10.0
  [77ba4419] NaNMath v1.1.3
  [bac558e1] OrderedCollections v1.8.1
  [90014a1f] PDMats v0.11.35
  [f03ddd90] Piecewise v0.1.1
  [21216c6a] Preferences v1.4.3
  [43287f4e] PtrArrays v1.3.0
  [1fd47b50] QuadGK v2.11.2
  [189a3867] Reexport v1.2.2
  [ae029012] Requires v1.3.1
  [79098fc4] Rmath v0.8.0
  [efcf1570] Setfield v1.1.2
  [a2af1166] SortingAlgorithms v1.2.1
  [276daf66] SpecialFunctions v2.5.1
  [1e83bf80] StaticArraysCore v1.4.3
  [82ae8749] StatsAPI v1.7.1
  [2913bbd2] StatsBase v0.34.5
  [4c63d2b9] StatsFuns v1.5.0
  [efe28fd5] OpenSpecFun_jll v0.5.6+0
  [f50d1b31] Rmath_jll v0.5.1+0
  [56f22d72] Artifacts
  [2a0f44e3] Base64
  [ade2ca70] Dates
  [8ba89e20] Distributed
  [9fa8497b] Future
  [b77e0a4c] InteractiveUtils
  [8f399da3] Libdl
  [37e2e46d] LinearAlgebra
  [56ddb016] Logging
  [d6f4376e] Markdown
  [de0858da] Printf
  [9a3f8284] Random
  [ea8e919c] SHA v0.7.0
  [9e88b42a] Serialization
  [6462fe0b] Sockets
  [2f01184e] SparseArrays v1.10.0
  [10745b16] Statistics v1.10.0
  [4607b0f0] SuiteSparse
  [fa267f1f] TOML v1.0.3
  [8dfed614] Test
  [cf7118a7] UUIDs
  [4ec0a83e] Unicode
  [e66e0078] CompilerSupportLibraries_jll v1.1.1+0
  [4536629a] OpenBLAS_jll v0.3.23+4
  [05823500] OpenLibm_jll v0.8.1+4
  [bea87d4a] SuiteSparse_jll v7.2.1+1
  [8e850b90] libblastrampoline_jll v5.11.0+0
Precompiling packages finished.
  17 dependencies successfully precompiled in 8 seconds. 49 already precompiled.
     Testing Running tests...
Test Summary: | Pass  Total  Time
Formula       |   98     98  0.1s
Test Summary: | Pass  Total  Time
Piece         |   16     16  0.1s
Test Summary:     | Pass  Total  Time
PiecewiseFunction |    8      8  0.1s
Test Summary: | Pass  Total  Time
POLY          |    5      5  0.1s
TAIL: Test Failed at ~/.julia/packages/Piecewise/6hjeI/test/formulas.jl:26
  Expression: Piecewise.TAIL_F(1, [1, 2, 0, 4, 5], 1) == 0.30268837405404053
   Evaluated: 0.3026883740540405 == 0.30268837405404053

Stacktrace:
 [1] macro expansion
   @ ~/.julia/.../share/julia/stdlib/v1.10/Test/src/Test.jl:672 [inlined]
 [2] macro expansion
   @ ~/.julia/packages/Piecewise/6hjeI/test/formulas.jl:26 [inlined]
 [3] macro expansion
   @ ~/.julia/.../share/julia/stdlib/v1.10/Test/src/Test.jl:1577 [inlined]
 [4] top-level scope
   @ ~/.julia/packages/Piecewise/6hjeI/test/formulas.jl:15
TAIL: Test Failed at ~/.julia/packages/Piecewise/6hjeI/test/formulas.jl:31
  Expression: Piecewise.TAIL_F(1, [1, 2, 3, 4, 5], 1) == 0.14729513605358416
   Evaluated: 0.14729513605358413 == 0.14729513605358416

Stacktrace:
 [1] macro expansion
   @ ~/.julia/.../share/julia/stdlib/v1.10/Test/src/Test.jl:672 [inlined]
 [2] macro expansion
   @ ~/.julia/packages/Piecewise/6hjeI/test/formulas.jl:31 [inlined]
 [3] macro expansion
   @ ~/.julia/.../share/julia/stdlib/v1.10/Test/src/Test.jl:1577 [inlined]
 [4] top-level scope
   @ ~/.julia/packages/Piecewise/6hjeI/test/formulas.jl:15
Test Summary: | Pass  Fail  Total  Time
TAIL          |   16     2     18  0.6s
ERROR: LoadError: Some tests did not pass: 16 passed, 2 failed, 0 errored, 0 broken.
in expression starting at ~/.julia/packages/Piecewise/6hjeI/test/formulas.jl:13
in expression starting at ~/.julia/packages/Piecewise/6hjeI/test/runtests.jl:5
ERROR: Package Piecewise errored during testing
```
