### 2025-07-25 ChristopheBerthod/Piecewise.jl#13 @ChristopheBerthod
These are inconsequential `float64` rounding errors, most likely due to different architectures. Directly comparing floats in the tests is not a good idea, indeed. Pending a better solution, I will replace the `==` by `≈`, which works with these two cases (for me at least):
```julia
julia> 0.3026883740540405 ≈ 0.30268837405404053
true
julia> 0.14729513605358413 ≈ 0.14729513605358416
true
```
