+++ Piecewise/test/formulas.jl
@@ -1,10 +1,10 @@
 
 @testset "POLY" begin
 
-    @test Piecewise.POLY_F(1, [1, 2, 3, 4, 5, 6]) == 21
+    @test Piecewise.POLY_F(1, [1, 2, 3, 4, 5, 6]) ≈ 21
     @test Piecewise.POLY_F([1, 2, 3], 2) == [2, 4, 6]
     @test Piecewise.POLY_F([1, 2, 3, 4]) == [1, -2, 3, -4]
-    @test Piecewise.POLY_F(1, [1, 2, 3, 4, 5, 6], 1) == 4.4071428571428575
+    @test Piecewise.POLY_F(1, [1, 2, 3, 4, 5, 6], 1) ≈ 4.4071428571428575
     @test typeof(POLY) == Formula
 
 end
@@ -12,23 +12,23 @@ end
 
 @testset "TAIL" begin
 
-    @test Piecewise.TAIL_F(1, [1, 2, 3, 4, 5]) == 0.25
+    @test Piecewise.TAIL_F(1, [1, 2, 3, 4, 5]) ≈ 0.25
     @test_throws ArgumentError Piecewise.TAIL_F([1, 2, -3, 4, 5], (0, 1), (true, true), true, true)
     @test Piecewise.TAIL_F([1, 2, -3, 4, 5], (0, 1), (true, true), true, false) == false
     @test Piecewise.TAIL_F([1, 2, 3, 4, 5], (0, 1), (true, true), true, true) == true
     @test Piecewise.TAIL_F([1, 2, -3, 4, 5], (0, 1), (true, true), true, false) == false
     @test Piecewise.TAIL_F([1, 2, 3, 4, 5], 2) == [2, 4, 3, 4, 5]
     @test Piecewise.TAIL_F([1, 2, 3, 4, 5]) == [1, -2, 3, -4, 5]
-    @test Piecewise.TAIL_F(1, [1, 2, 0, 0, 5], 0) == -0.2
-    @test Piecewise.TAIL_F(1, [1, 2, 0, 0, 5], 1) == 0.4
-    @test Piecewise.TAIL_F(1, [1, 2, 0, 0, 5], 5) == 0.13
-    @test Piecewise.TAIL_F(1, [1, 2, 0, 4, 5], 0) == 0.32958368660043297
-    @test Piecewise.TAIL_F(1, [1, 2, 0, 4, 5], 1) == 0.30268837405404053
-    @test Piecewise.TAIL_F(1, [1, 2, 3, 4, 0], 1) == 0.2044341744113004
-    @test Piecewise.TAIL_F(1, [1, 2, 3, 0, 0], 1) == 0.38888888888888884
-    @test Piecewise.TAIL_F(1, [1, 2, 3, 4, 0], 1) == 0.2044341744113004
-    @test Piecewise.TAIL_F(1, [1, 2, 1, 2, 1], 1) == 0.4205584583201638
-    @test Piecewise.TAIL_F(1, [1, 2, 3, 4, 5], 1) == 0.14729513605358416
+    @test Piecewise.TAIL_F(1, [1, 2, 0, 0, 5], 0) ≈ -0.2
+    @test Piecewise.TAIL_F(1, [1, 2, 0, 0, 5], 1) ≈ 0.4
+    @test Piecewise.TAIL_F(1, [1, 2, 0, 0, 5], 5) ≈ 0.13
+    @test Piecewise.TAIL_F(1, [1, 2, 0, 4, 5], 0) ≈ 0.32958368660043297
+    @test Piecewise.TAIL_F(1, [1, 2, 0, 4, 5], 1) ≈ 0.30268837405404053
[…]
