+++ PiecewiseLorentz/Project.toml
@@ -1,7 +1,7 @@
 name = "PiecewiseLorentz"
 uuid = "5265c256-a0b7-4541-80b8-9bb011dcc609"
 authors = ["Christophe Berthod <christophe.berthod@unige.ch>"]
-version = "0.1.1"
+version = "0.1.2"
 
 [deps]
 HypergeometricFunctions = "34004b35-14d8-5ef3-9330-4cdb6864b03a"
