### 2025-07-25 ChristopheBerthod/Piecewise.jl#13 @ChristopheBerthod
Version 0.1.2 has no direct comparison of `float64` and should pass all the tests.
