### 2025-07-29 ChristopheBerthod/Piecewise.jl#13 @ChristopheBerthod
@ranocha Did you get a chance to test the updated version?
