### 2025-08-19 ChristopheBerthod/Piecewise.jl#13 @ranocha
Please excuse the delayed response - vacation time...

Thank you for updating the tests. I can confirm that all tests pass locally for me.
