### 2025-08-19 ChristopheBerthod/Piecewise.jl#18 @ranocha
Automated tests
This is a part of https://github.com/openjournals/joss-reviews/issues/8329

Thanks for your package and for fixing the tests. Please excuse the delayed response - vacation time...

When continuing the review for JOSS, I noticed that automated tests are disabled (see the banner at the top of https://github.com/ChristopheBerthod/Piecewise.jl/actions/workflows/CI.yml) - same for documentation builds (https://github.com/ChristopheBerthod/Piecewise.jl/actions/workflows/Documenter.yml).

The documentation CI job did not run on the latest commit to `main`.

Could you please enable these CI jobs again and demonstrate that they work in a commit to the `main` branch so that I can check the corresponding item in the JOSS review list? You can consider doing this after merging https://github.com/ChristopheBerthod/Piecewise.jl/pull/17, for example.
