### 2025-08-19 ChristopheBerthod/Piecewise.jl#18 @ChristopheBerthod
Thanks a lot for the spellchecker. The CI and documentation jobs are enabled again and ran successfully after merging [#17](https://github.com/ChristopheBerthod/Piecewise.jl/pull/17).
