### 2025-08-19 ChristopheBerthod/Piecewise.jl#18 @ranocha
Thanks!
