### 2025-09-01 9b09bdf
Update paper.md
