### 2025-09-01 #8329 @ChristopheBerthod
All boxes can be ticked (I don't seem to have permissions for doing it).

Version number: v0.1.2
Zenodo DOI: 10.5281/zenodo.17016438
