### 2025-09-22 #8329 @gkthiruvathukal
@danielskatz  Hey, thanks for the reminder. By the way, is this reminder feature new? I feel so dumb for just noticing it now. This would really help me to keep up on submissions where I'm waiting on reviewers, especially. (And perhaps it would help them, too.)
