### 2025-09-22 #8329 @danielskatz
@ChristopheBerthod - I think there's just one small change needed in the paper, as shown in https://github.com/ChristopheBerthod/Piecewise.jl/pull/19.

Please merge this, or let me know if you disagree, then we can proceed.

Also, let me know if you have any other issues with the proof.
