### 2025-09-22 #8329 @ChristopheBerthod
Thank you. I went through the proofs and found only one issue:

One of the references that was a preprint has meanwhile appeared in print. "arXiv:2504.02739" with DOI link https://doi.org/10.48550/arXiv.2504.02739 could be replaced by "Physical Review Research, 7, 033038" with DOI link https://doi.org/10.1103/nzrk-yfqk. Should I update the paper myself?
