+++ joss/paper.bib
@@ -22,9 +22,11 @@
 @article{Morpurgo-2025,
   author    = {Morpurgo, Giacomo and Berthod, Christophe and Giamarchi, Thierry},
   title     = {Universal low-density power laws of the dc conductivity and {Hall} constant in the self-consistent {Born} approximation},
-  journal   = {arXiv:2504.02739},
+  journal   = {Physical Review Research},
   year      = {2025},
-  doi       = {10.48550/arXiv.2504.02739}
+  volume    = {7},
+  pages     = {033038},
+  doi       = {10.1103/nzrk-yfqk}
 }
 
 @misc{MagnetoTransport.jl,
